requirements.txt
SpeechRecognition
pyttsx3
paho-mqtt
pyaudio
